Revamp of shell file precedence
Should now play nicer with Linux graphical shells (as opposed to the macOS
terminal which is always login). In short, change so that more configuration
goes into .profile and .bashrc, and less into .bash_profile. This is because
we can't guarantee that .bash_profile does get sourced, and we don't need
e.g. prompt variables cluttering up the environment.