Copy conda config to .bashrc
The `conda` command is a shell function, not a file! That means it is not
inherited and needs conda.sh to be run for every shell that needs it.
  We can source conda.sh from both .profile and .bashrc, thereby covering
login and interactive shells. However, this doesn't solve the issue: the
conda command is still unavailable to non-login, non-interactive shells.
  For example, modifying the code from https://askubuntu.com/a/879400,
we have:

@ Interactive, non-login shell. Regular terminal
echo $-; shopt login_shell; which python; conda info --envs
@ Interactive login shell
bash -l
echo $-; shopt login_shell; which python; conda info --envs
@ Non-interactive, non-login shell
ssh localhost 'echo $-; shopt login_shell; which python; conda info --envs'
@ Non-interactive login shell
echo 'echo $-; shopt login_shell; which python; conda info --envs' | ssh localhost

  It is the penultimate one of these that is problematic - as above, not all
shells have conda available to them... There is no easy way around this
so, for now, we source conda.sh twice: in .profile (where we also activate
the default env) and in .bashrc. Non-interactive, non-login shells have
to rely on the conda environment of the process that spawned them.