Explicit checks for variables set in .machine_specific_config
Currently just test if the file we want to source exists, but this is
fairly course. Now check if the variable is set, and if it is but we
still can't find the file, then we print a warning.
  Also tweaked the CONDA_DEFAULT_ENV behaviour slightly as activating
conda by default can be problematic in some contexts. Now, conda is only
activated in .profile if the default env has a value, as `conda activate`
is equivalent to `conda activate base`; in other words, conda was getting
activated even without an explicit default env.